Adding opt pass to transform printf calls into translator-friendly calls.
This pass was designed to work in the first opt flow. It searches
for calls to the printf() function, analyzes its arguments and
breaks this printf() call into many printf() calls with one argument
each. If one of the operands is a single or double floating point,
it adds a padding zeroed integer before it in the arguments list.

Updating static-bt CMakeLists.txt to be more flexible with the include dir

Making PrintfBreak pass understand double percentage in printfs