Making RelocationReader propagate unresolved relocations
This commit changes OiInstTranslation and RelocationReader logic to
generate a reference to a symbol whenever the location of the
symbol couldn't be determined. This happens when the used symbol is
not in the program, but in the libc library. Previously, all such
requests were either ignored (for objects) or processed via the
libc call translation mechanism (for functions).