[static-bt] Teach static-bt to use information from IJMP insns
In a past commit, OpenISA backend was modified to include new instructions
(IJMPHI and IJMP) dedicated to encode an indirect jump. This instruction
makes explicit the location of jump tables, separating the base address
from the index used to select which jump location to jump to. This commit
modifies the static translator to use the information from IJMP
instructions, helping it to clearly identify the targets of indirect jumps
and improve the precision of the translated code by avoiding conservatively
creating unnecessary edges in the CFG (which was done before).