[static bt] When translating IJMPs, emit SwitchInst instead of IndirectBrs
The use of IndirectBr instructions in LLVM IR code as a result of translation
is causing problems when using opt. In SPEC programs, opt frequently hits
assertions complaining it cannot split a critical edge in an IndirectBr
instruction. Since the LLVM infrastructure seems to be better prepared to
handle SwitchInsts instead of IndirectBrs, we use the former.