[static bt] Refine addresses used in indirect calls in oneregion mode
Previously, whenever an indirect call was found, in oneregion mode it was
translated to an indirectbr with all functions in the binary as possible
targets (BB successors). This patch refines this set to use only addresses
of functions that we know can be target of indirect calls. We discover this
via relocations against the .text section that target a function. If such
relocations are present, these functions are considered as possible targets
to indirect calls and this is appropriately used when creating the indirectbr
instruction to call as well as the indirectbr instruction used to return from
regular functions.