[static bt] Make returns in oneregion do not depend on hashtable lookups
Oneregion depended on looking up a hashtable to convert guest addresses to host
addresses whenever generating code to return from a function. However, this is
unnecessary in a purely static translation and was originally done to evaluate
the overhead in a dynamic bt system. This commit removes this hashtable lookup
in favor of directly referencing basic blocks in the host code. Since hashtable
creation auxiliary functions now have 0 users, they were removed as well.