[static bt] Adding the new mode abi-locals
Previously this was maintained as a separate branch, but now it is
a new option in the regular static-bt. This mode reduces the number
of registers synced when entering or exiting a translation region,
that is, a function. It only synchronizes registers that are expected
to carry useful information from one function to another as per the
ABI.