[static-bt] Fix abi-locals mode saved regs range
In double precision contexts, the abi says register D1 can also be used
to pass registers or arguments. Fix this to make 482.sphinx3 work again.
Also add T1 in integer contexts to the list of synced registers, same
reason. Change the total number of double registers saved to 32 instead
of 64 because currently we are not using all 64 double regs.