[static-bt] Fix ctype_bloc family of functions at syscalls iface
Clean up code that previously created a loop to convert many pointers
related to the return of ctype_bloc family of functions. Only one
pointer needs to be converted. Change the strategy to copy the
converted pointer to global memory instead of patching the value
pointed to by the return of the original bloc function. Since this
original pointer can point to read-only memory, binaries were
segfaulting in ARM.