[static bt] Fix abi-locals usage of count operand
During a function call, according to the ABI, some integer regs
may be 'shadowed' by a double register. Therefore, if a call says
it uses 2 parameters, if the first param is a double and the second
is an integer, the two first int regs are discarded and the second
param is encoded in the third reg. This patch makes abi-locals
logic aware of this.